chore: refactor classes to functions
